refactor: RepositoryProvider 이름 변경
