+++ app/src/main/java/woowacourse/shopping/AlarmScheduler.kt
@@ -4,13 +4,13 @@ import android.app.AlarmManager
 import android.app.PendingIntent
 import android.content.Context
 import android.content.Intent
-import woowacourse.shopping.data.source.local.notification.NotificationDataSource
+import woowacourse.shopping.data.source.local.notification.NotificationSettingDataSource
 
 class AlarmScheduler(
     private val context: Context,
     private val requestCode: Int,
     private val receiver: Class<*>,
-    private val notificationDataSource: NotificationDataSource,
+    private val notificationSettingDataSource: NotificationSettingDataSource,
 ) {
     private val alarmManager = context.getSystemService(Context.ALARM_SERVICE) as AlarmManager
 
@@ -24,7 +24,7 @@ class AlarmScheduler(
     }
 
     fun schedule(delayMillis: Long) {
-        if (!notificationDataSource.isNotificationEnabled()) return
+        if (!notificationSettingDataSource.isNotificationEnabled()) return
         alarmManager.set(
             AlarmManager.RTC_WAKEUP,
             System.currentTimeMillis() + delayMillis,
