+++ app/src/main/java/woowacourse/shopping/ShoppingApplication.kt
@@ -7,7 +7,7 @@ import android.content.Context
 import androidx.datastore.core.DataStore
 import androidx.datastore.preferences.core.Preferences
 import androidx.datastore.preferences.preferencesDataStore
-import woowacourse.shopping.di.RepositoryProvider
+import woowacourse.shopping.di.AppContainer
 
 val Context.dataStore: DataStore<Preferences> by preferencesDataStore(name = "auth_token")
 
@@ -15,7 +15,7 @@ class ShoppingApplication : Application() {
     override fun onCreate() {
         super.onCreate()
         createAlarmChannel()
-        RepositoryProvider.init(
+        AppContainer.init(
             context = this,
             id = "aksworns22",
             password = "password",
