+++ app/src/main/java/woowacourse/shopping/presentation/order/OrderScreen.kt
@@ -54,7 +54,6 @@ fun OrderScreen(
     productIds: List<Long>,
     onBackClick: () -> Unit,
     onOrderSuccess: () -> Unit,
-    onEnterOrder: () -> Unit,
     modifier: Modifier = Modifier,
     viewModel: OrderViewModel = viewModel(),
 ) {
@@ -63,7 +62,6 @@ fun OrderScreen(
     val context = LocalContext.current
 
     LaunchedEffect(Unit) {
-        onEnterOrder()
         viewModel.initializePaymentItems(productIds)
         lifecycleOwner.repeatOnLifecycle(Lifecycle.State.STARTED) {
             viewModel.event.collect { event ->
